Disable `cgo` (#700)
It's possible that the tool is being used in environments where libc is
unavailable or there is a version difference to what was available during
compilation, for example CI environments.

Disaling the use of `cgo` will allow the tool to be more portable, as
the dependency on libc is removed.

Co-authored-by: Hidetake Iwata <int128@gmail.com>